requirements.txt
streamlit==1.33.0
scikit-learn==1.3.2
pandas==2.2.1
nltk==3.8.1
pillow==10.3.0
pytesseract==0.3.10
matplotlib==3.8.4
seaborn==0.13.2
